+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ### Added
 * Add prime number caching [#5](https://github.com/ddcprg/prime-number-service/issues/5).
 * Build Docker image [#6](https://github.com/ddcprg/prime-number-service/issues/6).
+### Fixed
+* More sensible metrics [#17](https://github.com/ddcprg/prime-number-service/issues/17).
 
 # [0.1.1] 2018-04-05
 ### Added